Added a new suite, queryCSS, i.e., query functions with styleWithCSS being set to true. This apparently makes a difference on Firefox. For now, this is otherwise a 1:1 clone of the other query tests, which seems like a waste. Need to investiate whether the expectations of query functions should be different between styleWithCss == false and true (e.g., give preference to HTML attributes in the former case and to CSS style values in the latter?)
Also:
.) cleaned up the code for color handling 
.) added support for the 'transparent' keyword
.) added support for alpha values
.) added preliminary support for % values in rgb() and rgba() notations

Currently, the support for % values is commented out since I suspect it'd lead to
false negative on comparisons. 

For both % values and the yet-to-be written hsl/hsla values I'll likely need to store
the values as-is and do exact matching on like-with-like comparisons, and fuzzy matching
over different original notations.