RichText 2 test suite:
Changed output to show both non-strict (verify HTML only) and strict (verify both HTML and selection) 
versions of tests (where applicable).
This artificially inflates the count of tests, but discussing it in group we felt that it's good
enough for a first draft, to function as basis for discussion. We can split it up again later if
necessary.

Other fixes:

.) Changed function-based tests to take a string that is run via eval(). This simplifies the conversion
between python and JS, and also allows us to output the function run in the reult table.

.) Fixed a bug when certain attributes (class, style and/or id) are faded because they are irrelevant
for the outcome. The old code did this in an incorrect order

.) Changed the code that produces inline selection markers to try and apply comment node children
for that purposes, and only fall back to use intra-element markers (e.g., "br |>") when this throws
an exception.

.) Added a global variable for selection, but this doesn't yet fully work in FF. Seems there are
problems with <iframe> in that regard.

.) Added a try-catch block around outputting a single test result in case something fails there
(which it really shouldn't), to avoid the whole suite come crashing down.

.) Changed all tests to specify a selection that so far didn't.